[#98]아기돼지뷰 재정리 완료
